Parameterise values, add "sudo" to cmds.sh

+++ aws-personal-vm/cmds.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -v
 
 # This script can be run ONLY IF the EFS is not attached to /data AND IF the EBS volume that is attached to / is not the one that is managed bv terraform
 # Most likely after the instance has been terminated due to changes in terraform configuration
@@ -18,6 +18,13 @@ device="/dev/sda1"
 
 sleep 120
 
+if [[ "$old_volume_id" == "$new_volume_id" ]]; then
+  echo "something went wrong - old_volume_id ${old_volume_id} is the same as new"
+  echo "restarting instance ${instance_id}"
+  aws ec2 start-instances --instance-ids "${instance_id}"
+  exit 1
+fi
+
 aws ec2 detach-volume --instance-id "${instance_id}" --volume-id "${old_volume_id}"
 aws ec2 delete-volume --volume-id "${old_volume_id}"
 aws ec2 attach-volume --instance-id "${instance_id}" --volume-id "${new_volume_id}" --device "${device}"
@@ -28,4 +35,4 @@ aws ec2 start-instances --instance-ids "${instance_id}"
 
 sleep 120
 
-ssh linux.norbert-logiewa.co.uk "mkdir -p $efs_mount_point && apt update && apt install -y git nfs-{common,kernel-server} curl wget vim sed && mount -t nfs -o nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport ${efs_id}.efs.${region}.amazonaws.com:/ $efs_mount_point"
+ssh linux.norbert-logiewa.co.uk "sudo mkdir -p $efs_mount_point && sudo apt update && sudo apt install -y git nfs-{common,kernel-server} curl wget vim sed && sudo mount -t nfs -o nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport ${efs_id}.efs.${region}.amazonaws.com:/ $efs_mount_point"
